move threadpool
fix cmake
